requirements.txt
absl-py==0.6.1
appdirs==1.4.3
appnope==0.1.2
asgiref==3.3.4
asn1crypto==0.24.0
astor==0.7.1
astroid==2.1.0
backcall==0.2.0
beautifulsoup4==4.7.1
bs4==0.0.1
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
click==7.1.2
cryptography==2.4.2
cssselect==1.0.3
cycler==0.10.0
decorator==4.4.2
distlib==0.3.1
Django==3.2
dohq-tfs==1.0.135
et-xmlfile==1.1.0
fake-useragent==0.1.11
filelock==3.0.12
Flask==1.1.2
gast==0.2.0
grpcio==1.16.1
h5py==2.8.0
idna==2.8
importlib-metadata==4.0.1
importlib-resources==5.1.2
ipykernel==5.4.3
ipython==7.16.1
ipython-genutils==0.2.0
isort==4.3.4
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.2
joblib==0.16.0
json2html==1.3.0
jupyter-client==6.1.11
jupyter-core==4.7.1
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
lxml==4.3.2
Markdown==3.0.1
MarkupSafe==1.1.1
matplotlib==3.2.1
mccabe==0.6.1
ntlm-auth==1.2.0
numpy==1.19.5
opencv-python==4.1.0.25
openpyxl==3.0.7
pandas==0.23.4
parse==1.11.1
parso==0.8.1
pbr==5.5.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
prompt-toolkit==3.0.16
protobuf==3.6.1
ptyprocess==0.7.0
pycparser==2.19
pyee==5.0.0
Pygments==2.8.0
pylint==2.2.2
pyparsing==2.3.0
pyppeteer==0.0.25
pyquery==1.4.0
python-dateutil==2.7.5
pytz==2018.7
pyzmq==22.0.3
requests==2.21.0
requests-html==0.10.0
requests-ntlm==1.1.0
scikit-learn==0.23.2
scipy==1.2.0
six==1.11.0
sklearn==0.0
soupsieve==1.8
sqlparse==0.4.1
stevedore==3.3.0
tensorboard==1.12.0
tensorflow==1.12.0
termcolor==1.1.0
threadpoolctl==2.1.0
tornado==6.1
tqdm==4.31.1
traitlets==4.3.3
typed-ast==1.1.0
typing-extensions==3.7.4.3
urllib3==1.24.1
virtualenv==20.4.4
virtualenv-clone==0.5.4
virtualenvwrapper==4.8.4
w3lib==1.20.0
wcwidth==0.2.5
websockets==7.0
Werkzeug==1.0.1
wrapt==1.10.11
xlwt==1.3.0
zipp==3.4.1
